requirements.txt
newspaper3k==0.2.8
pytz==2023.3
requests==2.31.0
zhipuai==1.0.7
